Test graphs

Graphs.Test is a module containing generic graph structures which make as few hypotheses as possible beyond the basic interface. It is not part of the public API and may break unexpectedly.

Index

Graphs.Test

Graphs.Test.GenericDiGraph

Graphs.Test.GenericEdge

Graphs.Test.GenericGraph

Graphs.Test.generic_graph

Full docs

Graphs.Test — Module

Graphs.Test

A module that provides utilities for testing functions that should work with any Graphs.AbstractGraph.

Graphs.Test.GenericDiGraph — Type

GenericDiGraph{T} <: Graphs.AbstractGraph{T}

A directed graph type that can be used to tests functions that relay on the Graphs.jl interface.

Graphs.Test.GenericEdge — Type

GenericEdge <: Graphs.AbstractEdge

An edge type that can be used to tests functions that relay on the Graphs.jl interface.

Graphs.Test.GenericGraph — Type

GenericGraph{T} <: Graphs.AbstractGraph{T}

An undirected graph type that can be used to tests functions that relay on the Graphs.jl interface.

Graphs.Test.generic_graph — Method

generic_graph(g::Union{SimpleGraph, SimpleDiGraph})

Return either a GenericGraph or GenericDiGraph that wraps a copy of g.