Spanning trees

Graphs.jl contains a few algorithms to compute minimum spanning trees.

Index

Graphs.boruvka_mst

Graphs.is_planar

Graphs.kruskal_mst

Graphs.planar_maximally_filtered_graph

Graphs.prim_mst

Full docs

Graphs.boruvka_mst — Function

boruvka_mst(g, distmx = weights(g); minimize = true)

Return a tuple (mst, weights) where mst is a vector of edges representing the optimum (minimum, by default) spanning tree of a connected, undirected graph g with optional matrix distmx that provides distinct edge weights, and weights is the sum of all the edges in the solution by using Boruvka's algorithm. The algorithm requires that all edges have different weights to correctly generate a minimum/maximum spanning tree

Optional Arguments

minimize=true: if set to false, calculate the maximum spanning tree.

Graphs.kruskal_mst — Function

kruskal_mst(g, distmx=weights(g); minimize=true)
kruskal_mst(g, weight_vector; minimize=true)

Return a vector of edges representing the minimum (by default) spanning tree of a connected, undirected graph g with optional distance matrix distmx using Kruskal's algorithm.

Alternative to the distance matrix distmx, one can pass a weight_vector with weights ordered as edges(g).

Optional Arguments

minimize=true: if set to false, calculate the maximum spanning tree.

Graphs.prim_mst — Function

prim_mst(g, distmx=weights(g))

Return a vector of edges representing the minimum spanning tree of a connected, undirected graph g with optional distance matrix distmx using Prim's algorithm. Return a vector of edges.

Graphs.is_planar — Method

is_planar(g)

Determines whether or not the graph g is planar.

Uses the left-right planarity test.

References

Brandes 2009

Graphs.planar_maximally_filtered_graph — Method

planar_maximally_filtered_graph(g)

Compute the Planar Maximally Filtered Graph (PMFG) of weighted graph g. Returns a SimpleGraph{eltype(g)}.

Examples

using Graphs, SimpleWeightedGraphs
N = 20
M = Symmetric(randn(N, N))
g = SimpleWeightedGraph(M) 
p_g = planar_maximally_filtered_graph(g)

References

Tuminello et al. 2005